requirements/base.txt
Django==2.2.10
Markdown==3.1.1
py-bcrypt==0.4
pytz==2019.3
sqlparse==0.3.0
gunicorn
python-dotenv
greenlet
gevent
geoip2
django-ipware
WeasyPrint
Unidecode
requests